+++ lending/src/codatlending/manage_data_pull_operations.py
@@ -51,7 +51,7 @@ class ManageDataPullOperations:
                 res.pull_operation = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
-        elif http_res.status_code in [401, 404, 429]:
+        elif http_res.status_code in [401, 402, 403, 404, 429, 500, 503]:
             if utils.match_content_type(content_type, 'application/json'):
                 out = utils.unmarshal_json(http_res.text, Optional[shared.ErrorMessage])
                 res.error_message = out
@@ -63,13 +63,7 @@ class ManageDataPullOperations:
     
     def list(self, request: operations.ListPullOperationsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListPullOperationsResponse:
         r"""List pull operations
-        The *List pull operations* endpoint returns a list of [pull operations](https://docs.codat.io/lending-api#/schemas/PullOperation) made by your client.
-
-        A [pull operation](https://docs.codat.io/lending-api#/schemas/PullOperation) is a request to retrieve a specific data type from an integration.
-
-        ### Tips and traps
-
-        - The *List pull operations* endpoint does not support querying the `isCompleted` property. You can filter failed pull operations by querying `status!=Complete&&status!=NotSupported` instead.
+        Gets the pull operation history (datasets) for a given company.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())
         
@@ -107,7 +101,7 @@ class ManageDataPullOperations:
                 res.pull_operations = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
-        elif http_res.status_code in [400, 401, 404, 429]:
+        elif http_res.status_code in [400, 401, 402, 403, 404, 429, 500, 503]:
             if utils.match_content_type(content_type, 'application/json'):
                 out = utils.unmarshal_json(http_res.text, Optional[shared.ErrorMessage])
                 res.error_message = out