+++ lending/src/codatlending/models/operations/get_commerce_customer_retention_metrics.py
@@ -38,7 +38,7 @@ class GetCommerceCustomerRetentionMetricsResponse:
     commerce_report: Optional[shared_commercereport.CommerceReport] = dataclasses.field(default=None)
     r"""OK"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""